Remove note header from playground
