--- codex-ui/src/vue/components/note-header/index.ts
@@ -1,3 +1,0 @@
-import NoteHeader from './NoteHeader.vue';
-
-export { NoteHeader };
